# openapi.yaml
openapi: 3.1.0
info:
  title: A2A Registry API
  version: 0.1.0
  description: Lease-based registration and discovery for A2A Agent Cards. This registry API is not an A2A task/message transport.
servers:
  - url: http://localhost:3003
paths:
  /v1/agents:
    post:
      summary: Register an agent
      security:
        - writeToken: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationInput'
      responses:
        '201':
          description: Registered. The lease token is returned only on creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
    get:
      summary: Discover active agents
      parameters:
        - { name: skill, in: query, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: capability, in: query, schema: { type: string } }
        - { name: protocolBinding, in: query, schema: { type: string } }
        - { name: name, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        '200':
          description: Matching unexpired registrations
          content:
            application/json:
              schema:
                type: object
                required: [agents, total, revision]
                properties:
                  agents: { type: array, items: { $ref: '#/components/schemas/RegisteredAgent' } }
                  total: { type: integer }
                  nextCursor: { type: string }
                  revision: { type: integer }
...
components:
  securitySchemes:
    writeToken:
      type: http
      scheme: bearer
    leaseToken:
      type: apiKey
      in: header
      name: X-Registry-Lease-Token
...
  responses:
    Problem:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    AgentCard:
      type: object
      description: A2A Agent Card. Unknown fields are preserved for forward compatibility.
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        version: { type: string }
        supportedInterfaces:
          type: array
          items:
            type: object
            properties:
              url: { type: string, format: uri }
              protocolBinding: { type: string }
              protocolVersion: { type: string }
        capabilities: { type: object, additionalProperties: true }
        skills: { type: array, items: { type: object, additionalProperties: true } }
      additionalProperties: true
    RegistrationInput:
      allOf:
        - $ref: '#/components/schemas/RegistrationInputWithoutId'
        - type: object
          required: [id]
          properties:
            id: { type: string, pattern: '^[a-zA-Z0-9][a-zA-Z0-9._:-]{0,127}$' }
    RegistrationInputWithoutId:
      type: object
      required: [agentCard]
      properties:
        endpoint: { type: string, format: uri }
        agentCard: { $ref: '#/components/schemas/AgentCard' }
        ttlSeconds: { type: integer, minimum: 1 }
        metadata: { type: object, additionalProperties: { type: string } }
    RegisteredAgent:
      type: object
      required: [id, name, endpoint, agentCard, ttlSeconds, registeredAt, updatedAt, lastSeen, expiresAt, metadata, revision]
      properties:
        id: { type: string }
        name: { type: string }
        endpoint: { type: string, format: uri }
        agentCard: { $ref: '#/components/schemas/AgentCard' }
        ttlSeconds: { type: integer }
        registeredAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastSeen: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: { type: string } }
        revision: { type: integer }
    RegistrationResponse:
      type: object
      required: [status, agent]
      properties:
        status: { type: string, enum: [registered, updated] }
        agent: { $ref: '#/components/schemas/RegisteredAgent' }
        leaseToken: { type: string, description: Returned only when a registration is created. }
    Problem:
      type: object
      required: [type, title, status, detail, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        requestId: { type: string }